feat(dm-icon-core): port compose + marks + filters — full pixel parity (M5.8/9/10)
The integration triad, completing the icon pixel core:
- compose/{mod,field}.rs: renderTile + composeTile + composeField, the 9
  top-level lanes + 6 field sub-lanes with read-only diagnostics, plate
  detection/backdrop-swap, silhouette dock shadow (f32-store/f64-accumulate
  blur), layered mono, inscribe/stretch. Absorbs slice.rs (deleted) — its
  helpers are the real compose helpers; render_slice_tile stays as the
  Spike-4 compat shim through the ported pipeline.
- marks/{mod,styles}.rs: the Mark trait + 7 styles (shadow/halo/ring/satin/
  arc/fold/glass), arrow glyph, real Win11 classic-arrow badge (thread-local,
  mirrors the TS module global), stamp/outside-distance. `x**2` ported as
  d*d (correctly-rounded pow(_,2) == the IEEE product — dodges the pow LUT drift).
- filters/{mod,glass,pixel,sticker}.rs: applyFilter, chamfer distance, gloss,
  liquid-glass, kawaii pixelate (Math.trunc casts), die-cut sticker.
- shape_mask offsets widened i32→f64 (TS `number`; stampMask passes floats).

Gate (xtask m5-pixels + tests/icon-parity/m5/cells.ts): the FULL corpus pixel
differential is BYTE-IDENTICAL — every tier-A master + tier-B style-matrix cell
(1248 total, all 9 lanes / 6 field lanes / 7 marks / 5 filters / real arrow /
7 presets) matches the TS oracle exactly (0 diff bytes / 327,155,712) with every
lane exact. Spike-4 tri-target still PASS through the compose-based slice
(native↔wasm 240/240, TS↔Rust 0 diff). 47 cargo tests pass, 0 warnings.

The full render_tile wasm export + wasm↔native CI is an M6 concern (needs the
Config ABI); the shared compose primitives are already wasm↔native proven via
Spike-4's render_slice_tile.

Refs ADR-0019 M5, docs/plans/2026-07-10-m5-icon-core.md.

+++ crates/dm-icon-core/src/raster.rs
@@ -136,13 +136,13 @@ pub fn shape_mask(
     shape: IconShape,
     buffer_size: usize,
     shape_size: usize,
-    offset_x: i32,
-    offset_y: i32,
+    offset_x: f64,
+    offset_y: f64,
 ) -> Vec<f64> {
     assert!(shape_size > 0, "shapeSize must be positive");
     let size = shape_size as f64;
-    let ox = offset_x as f64;
-    let oy = offset_y as f64;
+    let ox = offset_x;
+    let oy = offset_y;
     let mut mask = vec![0.0f64; buffer_size * buffer_size];
     let grid_w = buffer_size + 1;
     let mut grid = vec![0u8; grid_w * grid_w];
@@ -398,7 +398,7 @@ mod tests {
 
     #[test]
     fn circle_mask_interior_exterior_and_conservation() {
-        let mask = shape_mask(IconShape::Circle, 64, 64, 0, 0);
+        let mask = shape_mask(IconShape::Circle, 64, 64, 0.0, 0.0);
         assert_eq!(mask[32 * 64 + 32], 1.0); // centre
         assert_eq!(mask[0], 0.0); // corner
         // every value in [0,1]
